Fix bearerFormat for djangorestframework-simplejwt
When using the `Authorization` header, the schema output should use the
`type: http` and `scheme: bearer` values with `bearerFormat` set to the
first value of `AUTH_HEADER_TYPES`.

The `drf_spectacular/contrib/rest_framework_jwt.JWTScheme` extension
takes this approach.

Currently if the first value of `AUTH_HEADER_TYPES` is `Bearer` the
value of `bearerFormat` is set to `JWT`, but if the value is `JWT` then
`type: apiKey` is used instead. This seems rather inconsistent and is
purely based on the order of items in `AUTH_HEADER_TYPES` which has no
bearing on the whether the prefix is acceptable. The first value is only
used in preference by `djangorestframework-simplejwt` for value to use
in the `WWW-Authenticate` header.