Fix to set exclusive{Maximum,Minimum} with DecimalValidator.
If we are have max digits set to 4 and decimal places set to 1, we
expect the maximum value to be 999.9, but we are setting it to be 1000.
In this case we expect the maximum bound to be exclusive, not inclusive
as is the default.

Obviously, if we encounter another non-DecimalValidator that would
override, we'd need to remove the exclusive flag.

Note: This will need tweaking for OpenAPI 3.1 which uses a newer version
of the JSON Schema specification. In that case `exclusiveMaximum` and
`exclusiveMinimum` are no longer boolean values and should replace the
`maximum` and `minimum` properties instead.