Provide all fixed field names for the components object
The OpenAPI components object (https://swagger.io/specification/\#components-object)
is used to store schema definitions that can be referred to using the
 pattern, encouraging re-use. drf-spectacular only uses the
'schemas' and 'securitySchemes' properties internally, but extension
developers may wish to use the remaining properties, so they are
provided as a convenience.

Note that the 'pathItems' property was added in OpenAPI 3.1.0.